pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cv-alignment-studio"
version = "0.1.0"
description = "Streamlit-powered CV/job description alignment platform."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
    { name = "CV Alignment Studio", email = "support@example.com" }
]
dependencies = [
    "pandas==2.3.3",
    "numpy==1.26.4",
    "scikit-learn==1.4.2",
    "python-docx==1.1.2",
    "pdfminer.six==20231228",
    "rapidfuzz==3.9.6",
    "pyyaml==6.0.2",
    "watchdog==4.0.1",
    "openai==1.51.0",
    "plotly==5.23.0",
    "streamlit==1.50.0",
    "sentry-sdk==2.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-cov==5.0.0",
    "mypy==1.11.2",
    "ruff==0.7.1",
    "types-PyYAML==6.0.12.20240808",
    "types-requests==2.32.0.20241016",
]

[project.urls]
Homepage = "https://example.com"
Repository = "https://example.com/repo"

[tool.setuptools]
py-modules = ["match_cvs", "app", "automation"]

[tool.setuptools.package-data]
"*" = ["skills.yaml"]

[tool.ruff]
target-version = "py311"
select = ["E", "F", "I", "UP", "PL"]
ignore = ["E501"]
fix = false

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_return_any = true
strict_optional = true
check_untyped_defs = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "pandas",
    "pandas.*",
    "numpy",
    "numpy.*",
    "plotly",
    "plotly.*",
    "streamlit",
    "streamlit.*",
    "sentry_sdk",
    "sentry_sdk.*",
    "sklearn",
    "sklearn.*",
    "watchdog",
    "watchdog.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "--strict-markers --disable-warnings"
testpaths = ["tests"]
